refactor(magic,combat): review cleanups — batch Audere queries, shared test fixture module (#873)
