+++ src/world/combat/tests/test_audere_cleanup.py
@@ -10,7 +10,7 @@ from world.combat.services import cleanup_completed_encounter
 from world.conditions.models import ConditionInstance
 from world.magic.audere import AUDERE_CONDITION_NAME, PendingAudereOffer, offer_audere
 from world.magic.factories import CharacterAnimaFactory, PendingAudereOfferFactory
-from world.magic.tests.test_audere_offer_surface import build_audere_gate_fixture
+from world.magic.tests.audere_test_helpers import build_audere_gate_fixture
 from world.mechanics.constants import EngagementType
 from world.mechanics.engagement import CharacterEngagement
